fix(repositories): change entity ID types from Long to Integer

+++ src/main/java/com/acd/researchrepo/repository/ResearchPaperRepository.java
@@ -4,6 +4,6 @@ import com.acd.researchrepo.model.ResearchPaper;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResearchPaperRepository extends JpaRepository<ResearchPaper, Long> {
+public interface ResearchPaperRepository extends JpaRepository<ResearchPaper, Integer> {
 
 }
